Setup bulk record inserting process
This process uses different new functions along the way:
* 'fct_bulk_insert_records', the main inserting handler. Checks the amount sums and inserts the records when no errors are reported
* 'fct_transform_records_input' to fetch the records from $_REQUEST and modify its structure to return a usable set of records data
* 'fct_sanitize_record_data' to detect missing and invalid input fields and do some input sanitization
* 'fct_get_required_record_fields' to return the required fields before inserting the record
* 'fct_bulk_insert_records_notices' parses the registered errors before inserting the records

Additional hooks and helper functions are defined in the admin area.

Though the process isn't yet finished and working perfect, this commit
should store the logic just for now. Further work is required.

+++ includes/admin/admin.php
@@ -163,17 +163,17 @@ class Fiscaat_Admin {
 	 */
 	public function dev_content() {
 
-		// Bail if no add_content query arg
+		// Update period account count
 		if ( isset( $_REQUEST['period_account_count'] ) && $_REQUEST['period_account_count'] ) {
 			fct_update_period_account_count( fct_get_current_period_id() );
 		}
 
-		// Bail if no add_content query arg
+		// Create initial content
 		if ( isset( $_REQUEST['add_content'] ) && $_REQUEST['add_content'] ) {
 			fct_create_initial_content();
 		}
 
-		// Bail if no del_content query arg
+		// Delete accounts and periods
 		if ( isset( $_REQUEST['del_content'] ) && $_REQUEST['del_content'] ) {
 
 			// Delete all accounts
@@ -183,6 +183,15 @@ class Fiscaat_Admin {
 
 			// Delete all periods
 			foreach ( get_posts( array( 'post_type' => fct_get_period_post_type(), 'fields' => 'ids', 'numberposts' => -1 ) ) as $post_id ) {
+				wp_delete_post( $post_id, true ); // force delete
+			}
+		}
+
+		// Delete records
+		if ( isset( $_REQUEST['del_records'] ) && $_REQUEST['del_records'] ) {
+
+			// Delete all accounts
+			foreach ( get_posts( array( 'post_type' => fct_get_record_post_type(), 'fields' => 'ids', 'numberposts' => -1 ) ) as $post_id ) {
 				wp_delete_post( $post_id, true ); // force delete
 			}
 		}
